Extracted Contact superclass
